perf: lazy load command palette dialog
